feature #2699: Fix index type in string loop
(cherry picked from commit e6fc3ec37a67ce0af6bc11e23db705c911b4b9d1)

+++ src/vm/VirtualMachine.cc
@@ -2692,7 +2692,7 @@ int VirtualMachine::generate_context(string &files, int &disk_id, string& token_
         files += files_ds;
     }
 
-    for (int i=0;i<files.length();i++)
+    for (size_t i=0;i<files.length();i++)
     {
         if (files[i] == '\n')
         {
